backend/requirements.txt
asgiref==3.8.1
black==25.1.0
click==8.2.1
colorama==0.4.6
Django==5.2.1
django-cors-headers==4.7.0
django-filter==25.1
django-role-permissions==3.2.0
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
drf-nested-routers==0.94.2
flake8==7.2.0
isort==6.0.1
mccabe==0.7.0
mypy_extensions==1.1.0
mysqlclient==2.2.7
packaging==25.0
pathspec==0.12.1
pillow==11.2.1
platformdirs==4.3.8
pycodestyle==2.13.0
pyflakes==3.3.2
PyJWT==2.9.0
python-decouple==3.8
python-dotenv==1.1.0
sqlparse==0.5.3
tzdata==2025.2
